changed visualisation util functions
