+++ utils_cv/tracking/plot.py
@@ -1,12 +1,129 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
+from collections import OrderedDict
 import os.path as osp
 from typing import Dict, List, Tuple
-import cv2
 import numpy as np
 
+import cv2
+import decord
+import io
+import IPython.display
+from PIL import Image
+from time import sleep
+
 from .bbox import TrackingBbox
+from .model import _get_frame
+
+def plot_single_frame(
+    results: Dict[int, List[TrackingBbox]], input_video: str, frame_id: int
+) -> None:
+    """ 
+    Plot the bounding box and id on a wanted frame. Display as image to front end. 
+
+    Args:
+        results: dictionary mapping frame id to a list of predicted TrackingBboxes
+        input_video: path to the input video
+        frame_id: frame_id for frame to show tracking result
+    """
+    results = OrderedDict(sorted(results.items()))
+    
+    # Assign bbox color per id
+    unique_ids = list(
+        set([bb.track_id for frame in results.values() for bb in frame])
+    )
+    color_map = assign_colors(unique_ids)
+
+    # Get frame from video
+    im = _get_frame(input_video, frame_id) 
+
+    # Extract tracking results for wanted frame, and draw bboxes+tracking id, display frame
+    cur_tracks = results[frame_id]
+    if len(cur_tracks) > 0:
+        im = draw_boxes(im, cur_tracks, color_map)    
+    im = Image.fromarray(im)    
+    IPython.display.display(im)
+
+def play_video(
+    results: Dict[int, List[TrackingBbox]], input_video: str
+) -> None:
+    """ 
+     Plot the predicted tracks on the input video. Displays to front-end as sequence of images stringed together in a video.
+
+    Args:
+        results: dictionary mapping frame id to a list of predicted TrackingBboxes
+        input_video: path to the input video        
+    """
+
+    results = OrderedDict(sorted(results.items()))
+
+    # assign bbox color per id
+    unique_ids = list(
+        set([bb.track_id for frame in results.values() for bb in frame])
+    )
+    color_map = assign_colors(unique_ids)
+
+    # read video and initialize new tracking video
+    video_reader = decord.VideoReader(input_video)
+
+    # set up ipython jupyter display
+    d_video = IPython.display.display("", display_id=1)
+
+    # Read each frame, add bbox+track id, display frame    
+    for frame_idx in range(len(results)):         
+        im = video_reader.next().asnumpy()        
+        cur_tracks = results[frame_idx]
+        if len(cur_tracks) > 0:
+            cur_image = draw_boxes(im, cur_tracks, color_map)
+
+        f = io.BytesIO()
+        im = Image.fromarray(im)
+        im.save(f, "jpeg")
+        d_video.update(IPython.display.Image(data=f.getvalue()))
+        sleep(0.000001)        
+
+def write_video(
+    results: Dict[int, List[TrackingBbox]], input_video: str, output_video: str
+) -> None:
+    """ 
+    Plot the predicted tracks on the input video. Write the output to {output_path}.
+
+    Args:
+        results: dictionary mapping frame id to a list of predicted TrackingBboxes
+        input_video: path to the input video
+        output_video: path to write out the output video
+    """
+    results = OrderedDict(sorted(results.items()))
+    # read video and initialize new tracking video
+    video = cv2.VideoCapture()
+    video.open(input_video)
+
+    image_width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
+    image_height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
+    fourcc = cv2.VideoWriter_fourcc(*"MP4V")
+    frame_rate = int(video.get(cv2.CAP_PROP_FPS))
+    writer = cv2.VideoWriter(
+        output_video, fourcc, frame_rate, (image_width, image_height)
+    )
+
+    # assign bbox color per id
+    unique_ids = list(
+        set([bb.track_id for frame in results.values() for bb in frame])
+    )
+    color_map = assign_colors(unique_ids)
+
+    # create images and add to video writer, adapted from https://github.com/ZQPei/deep_sort_pytorch
+    frame_idx = 0
+    while video.grab():
+        _, im = video.retrieve()
+        cur_tracks = results[frame_idx]
+        if len(cur_tracks) > 0:
+            im = draw_boxes(im, cur_tracks, color_map)
+        writer.write(im)
+        frame_idx += 1
+
+    print(f"Output saved to {output_video}.")
 
 
 def draw_boxes(
@@ -32,7 +149,7 @@ def draw_boxes(
         bottom = round(bb.bottom)
 
         # box text and bar
-        color = color_map[label]
+        color = color_map[label]        
         label = str(label)
 
         # last two args of getTextSize() are font_scale and thickness
@@ -65,7 +182,7 @@ def assign_colors(id_list: List[int],) -> Dict[int, Tuple[int, int, int]]:
 
     # adapted from https://github.com/ZQPei/deep_sort_pytorch
     for i in id_list2:
-        color = [int((p * ((i + 1) ** 5 - i + 1)) % 255) for p in palette]
+        color = [int((p * ((i + 1) ** 4 - i + 1)) % 255) for p in palette]
         color_list.append(tuple(color))
 
     color_map = dict(zip(id_list, color_list))
